+++ src/full/Agda/TypeChecking/Coverage/Cubical.hs
@@ -10,9 +10,9 @@ import qualified Data.Set as Set
 import Data.IntSet (IntSet)
 import qualified Data.IntSet as IntSet
 
-import Agda.Syntax.Common
+import Agda.Syntax.Common hiding (DataOrRecord)
 import Agda.Syntax.Position
-import Agda.Syntax.Internal hiding (DataOrRecord)
+import Agda.Syntax.Internal
 import Agda.Syntax.Internal.Pattern
 import Agda.Syntax.Common.Pretty (prettyShow)