+++ test/test-suite-covers-errors.sh
@@ -41,6 +41,14 @@ cat >> $ERRORS <<EOF
 NeedOptionRewriting
 EOF
 
+# Errors which became impossible for ordinary data definitions
+# but may be still be triggerable by reflection.
+#
+cat >> $ERRORS <<EOF
+UnexpectedModalityAnnotationInParameter
+UnexpectedTypeSignatureForParameter
+EOF
+
 # Errors we currently do not cover by the testsuite (TODO!).
 #
 cat >> $ERRORS <<EOF